reorganize code to make it easy to reuse payer in additional commands
